add models for detailed item view / Universalis

+++ app/src/main/java/com/example/frivolity/network/models/universalisapi/ApiItemSaleData.kt
@@ -2,7 +2,7 @@ package com.example.frivolity.network.models.universalisapi
 
 data class ApiItemSaleData(
     val hq: Boolean,
-    val pricePerUnit: Long,
+    val pricePerUnit: Int,
     val quantity: Int,
     val timestamp: Long,
     val onMannequin: Boolean,
